+++ apps/metadata-service-api/src/api/users/user.service.ts
@@ -5,6 +5,11 @@ const getUserTokens = async (chain: SupportedChain, account: string, collection?
   return external.reservoirApi.requestUserTokens(chain, account, collection, continuation);
 };
 
+const getUserCollections = async (chain: SupportedChain, account: string, collection?: string, name?: string) => {
+  return external.reservoirApi.requestUserCollections(chain, account, collection, name);
+};
+
 export default {
   getUserTokens,
+  getUserCollections,
 };